# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Go Image Converter API",
        "contact": {}
    },
    "paths": {
        "/compress_image": {
            "post": {
                "description": "Compress image with specified compression level (1-5)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Compress image",
                "operationId": "compress_image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "compression level (1-5)",
                        "name": "compression_level",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/convert_png_to_jpeg": {
            "post": {
                "description": "Convert image from PNG format to JPEG format",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Convert PNG to JPEG",
                "operationId": "convert_png_to_jpeg",
                "parameters": [
                    {
                        "type": "file",
                        "description": "image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/resize_image": {
            "post": {
                "description": "Resize image to the specified width and height",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Resize image",
                "operationId": "resize_image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "width",
                        "name": "width",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "height",
                        "name": "height",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}